Made some static integer casts in the center-line draw function to silence compiler about cast warnings.

+++ src/center_line.cpp
@@ -10,8 +10,8 @@ CenterLine::CenterLine(int x, int y, int width, int height) : mRect({ x, y, widt
 void CenterLine::onDraw(SDL_Renderer& renderer)
 {
   // draw the center line by forming a number squares as required.
-  for (float y = mRect.y; y < mRect.h; y += (1.93 * mRect.w)) {
-    SDL_Rect rect{ mRect.x, y, mRect.w, mRect.w };
+  for (float y = static_cast<float>(mRect.y); y < mRect.h; y += (1.93f * mRect.w)) {
+    SDL_Rect rect{ mRect.x, static_cast<int>(y), mRect.w, mRect.w };
     SDL_RenderFillRect(&renderer, &rect);
   }
 }
